deployment/requirements.txt
ultralytics
# torch==1.13.1
# torchvision==0.14.1
# opencv-python==4.6.0.66
# numpy==1.24.2
Flask==2.2.3
easyocr==1.6.2
onnx==1.12.0
onnxruntime-gpu==1.12.1
